code-mentor-server/requirements.txt
fastapi
uvicorn
python-jose
PyJWT
passlib[bcrypt]
python-dotenv
